% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoxgbResult.R
\docType{data}
\name{AutoxgbResult}
\alias{AutoxgbResult}
\title{Result of an autoxgboost call.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AutoxgbResult
}
\description{
Result of \code{\link{autoxgboost}}. A R6 class containing the following elements:
\describe{
  \item{optim_result [\code{\link[mlrMBO]{MBOSingleObjResult}}]}{Optimization result object. See: \code{\link[mlrMBO]{MBOSingleObjResult}}}
  \item{final_learner [\code{\link[mlr]{Learner}}]}{Xgboost learner with best found hyper paramater configuration.}
  \item{final_model [\code{\link[mlr]{WrappedModel}} | \code{NULL}]}{If \code{build.final.model=TRUE} in \code{\link{autoxgboost}} a \pkg{mlr} model build by the full dataset and \code{final.learner}.}
  \item{measures [\code{\link[mlr]{Measure}}]}{Measure used for optimization.}
}
And the following functions:\cr
\describe{
  \item{.$predict(newdata)}
  \item{.$get_tune_results()}
  \item{.$plot_pareto_front()}
}
}
\keyword{datasets}
