% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbOptimizer.R
\docType{data}
\name{AxgbOptimizer}
\alias{AxgbOptimizer}
\title{AutoxgboostMC Optimizer Class}
\format{[R6::R6Class]}
\usage{
AxgbOptimizer
}
\description{
AutoxgboostMC Optimizer Class
}
\seealso{
\code{\link{AxgbOptimizerSMBO}}
}
\keyword{datasets}
