% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbOptimizer.R
\docType{data}
\name{AxgbOptimizerSMBO}
\alias{AxgbOptimizerSMBO}
\title{Optimize using SMBO}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxgbOptimizerSMBO
}
\description{
Optimize using SMBO
}
\keyword{datasets}
