% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbPipeline.R
\docType{data}
\name{AxgbPipeline}
\alias{AxgbPipeline}
\title{Abstract Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxgbPipeline
}
\description{
Abstract Base Class
}
\seealso{
\code{\link{AxgbPipelineXGB}}
}
\keyword{datasets}
