% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbPipelineBuilder.R
\docType{data}
\name{AxgbPipelineBuilder}
\alias{AxgbPipelineBuilder}
\title{Abstract Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxgbPipelineBuilder
}
\description{
Abstract Base Class
}
\seealso{
\code{\link{AxgbPipelineBuilderXGB}}
}
\keyword{datasets}
