% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbPipelineBuilder.R
\docType{data}
\name{AxgbPipelineBuilderXGB}
\alias{AxgbPipelineBuilderXGB}
\title{Build a xgboost baselearner and Preproc Pipeline}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxgbPipelineBuilderXGB
}
\description{
Build a xgboost baselearner and Preproc Pipeline
}
\keyword{datasets}
