% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbPipeline.R
\docType{data}
\name{AxgbPipelineXGB}
\alias{AxgbPipelineXGB}
\title{Build a xgboost baselearner and Preproc Pipeline}
\format{[R6::R6Class] object inheriting from [AxgbPipeline].}
\arguments{
\item{early_stopping_measure}{[\code{\link[mlr]{Measure}}]\cr
Performance measure used for early stopping. Picks the first measure
defined in measures by default.}

\item{early_stopping_rounds}{[\code{integer(1L}]\cr
After how many iterations without an improvement in the boosting OOB error should be stopped?
Default is \code{10}.}

\item{early_stopping_fraction}{[\code{numeric(1)}]\cr
What fraction of the data should be used for early stopping (i.e. as a validation set).
Default is \code{4/5}.}

\item{impact_encoding_boundary}{[\code{integer(1)}]\cr
Defines the threshold on how factor variables are handled. Factors with more levels than the \code{"impact_encoding_boundary"} get impact encoded while factor variables with less or equal levels than the \code{"impact_encoding_boundary"} get dummy encoded.
For \code{impact_encoding_boundary = 0L}, all factor variables get impact encoded while for \code{impact_encoding_boundary = .Machine$integer.max}, all of them get dummy encoded.
Default is \code{10}.}

\item{tune_threshold}{[logical(1)]\cr
Should thresholds be tuned? This has only an effect for classification, see \code{\link[mlr]{tuneThreshold}}.
Default is \code{TRUE}.}

\item{max_nrounds}{[\code{integer(1)}]\cr
Maximum number of allowed boosting iterations. Default is \code{3000}.}
}
\description{
Build a xgboost baselearner and Preproc Pipeline
}
\section{Construction}{
\cr
 ```
 axgb = AxgbPipelineXGB$new()
 ```

Additional arguments that control the Pipeline can be set via Active Bindings.
}

\keyword{datasets}
