% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{fairf1}
\alias{fairf1}
\title{Variance of F1 Scores between groups}
\format{An object of class \code{Measure} of length 10.}
\usage{
fairf1
}
\description{
Variance of F1 Scores between groups
}
\keyword{datasets}
