% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{fairness_measures}
\alias{fairness_measures}
\title{Fairness Measures}
\arguments{
\item{grouping}{[`function` | `factor` | `character`]\cr
Either a function(df), a factor or a character column name that returns a factor.
If `function`, df is the output of `getTaskData()`.
This factor is used to split the data and compute differences between groups.}
}
\description{
Fair algorithms usually are defined as having similar performances within
different dataset splits.
We implement the following measures:\cr
}
\details{
Independence `fairpr` \cr
Sufficiency `fairf1` \cr
Callibration `fairppv` \cr

The arguments can be set via `extra.args` for all measures.
}
\examples{
\dontrun{setMeasurePars(fairf1, grouping = function(df) {as.factor(df$age > 30)})}
}
