% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{fairppv}
\alias{fairppv}
\title{Absolute differences of Positive Predictive Value between groups}
\format{An object of class \code{Measure} of length 10.}
\usage{
fairppv
}
\description{
Absolute differences of Positive Predictive Value between groups
}
\keyword{datasets}
