% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{fairpr}
\alias{fairpr}
\title{Absolute differences of Positive Rate between groups}
\format{An object of class \code{Measure} of length 10.}
\usage{
fairpr
}
\description{
Absolute differences of Positive Rate between groups
}
\keyword{datasets}
