% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{get_grouping}
\alias{get_grouping}
\title{Obtain a grouping factor from either a data column, an additional factor or a function.}
\usage{
get_grouping(extra_args, n_levels = NULL)
}
\description{
Obtain a grouping factor from either a data column, an additional factor or a function.
}
