% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axgb_result.R
\name{get_pareto_data_nonconvex}
\alias{get_pareto_data_nonconvex}
\title{Obtain the data required to plot non-convex pareto fronts}
\usage{
get_pareto_data_nonconvex(front_data, x, y, x_minimize)
}
\description{
Obtain the data required to plot non-convex pareto fronts
}
