% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{inject_noise_single_feature}
\alias{inject_noise_single_feature}
\title{Inject noise into a single feature of a dataset.}
\usage{
inject_noise_single_feature(task, feature, eps = 0.05)
}
\description{
Inject noise into a single feature of a dataset.
}
