% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{inject_noise_task}
\alias{inject_noise_task}
\title{Inject noise into every feature of a dataset.}
\usage{
inject_noise_task(task, eps = 0.01)
}
\description{
Inject noise into every feature of a dataset.
}
