% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{interpias}
\alias{interpias}
\title{Interaction Strength
For more info see Molnar et al., 2019 https://arxiv.org/abs/1904.03867}
\format{An object of class \code{Measure} of length 10.}
\usage{
interpias
}
\description{
Interaction Strength
For more info see Molnar et al., 2019 https://arxiv.org/abs/1904.03867
}
\keyword{datasets}
