% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{interpmec}
\alias{interpmec}
\title{Curve Complexity
For more info see Molnar et al., 2019 https://arxiv.org/abs/1904.03867}
\format{An object of class \code{Measure} of length 10.}
\usage{
interpmec
}
\description{
Curve Complexity
For more info see Molnar et al., 2019 https://arxiv.org/abs/1904.03867
}
\keyword{datasets}
