% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{interpnf}
\alias{interpnf}
\title{Number of features
See `?interpretability_measures` for additonal info
For more info see Molnar et al., 2019 https://arxiv.org/abs/1904.03867}
\format{An object of class \code{Measure} of length 10.}
\usage{
interpnf
}
\description{
Number of features
See `?interpretability_measures` for additonal info
For more info see Molnar et al., 2019 https://arxiv.org/abs/1904.03867
}
\keyword{datasets}
