% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{interpnf2}
\alias{interpnf2}
\title{Number of features via noise injection}
\format{An object of class \code{Measure} of length 10.}
\usage{
interpnf2
}
\arguments{
\item{eps}{[`numeric(1)`]\cr
Magnitude of injected noise. Default: 0.7}

\item{n}{[`integer(1)`]\cr
Repliactions for the corruption process. Results are averaged. Default: 1L.}
}
\description{
Creates number of features copies of the prediction data.
For each copy, noise is injected into a single feature.
Then we check how often the prediction changes with respect to the
unmodified data.
}
\keyword{datasets}
