% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{interpretability_measures}
\alias{interpretability_measures}
\title{Interpretability Measures}
\arguments{
\item{grid.size}{[`integer(1)`]\cr
Controls the grid to evaluate on. Default: 10L.}

\item{max_seg_cat}{[`integer(1)`]\cr
Max. number of segments for categorical variables. Default: 5L.}

\item{max_seg_num}{[`integer(1)`]\cr
Max. number of segments for numerical variables. Default: 5L.}

\item{epsilon}{[`numeric(1)`]\cr
1 - Minimum R^2.}
}
\description{
Several measures for interpretability have been defined in [Molnar et al., 2019](https://arxiv.org/abs/1904.03867).
We implement those measures the following measures:
}
\details{
Curve Complexity `interpmec` \cr
Interaction Strength `interpias` \cr
Number of features `interpnf` \cr

The arguments can be controlled via `extra.args` for the aforementioned measures.
}
\examples{
\dontrun{setMeasurePars(interpmec, grid.size = 10L)}
}
