% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axgb_result.R
\name{plot_parallel_coordinates}
\alias{plot_parallel_coordinates}
\title{Parallel Coordinates Plot}
\usage{
plot_parallel_coordinates(trim = 20L, plotly = FALSE)
}
\description{
Parallel Coordinates Plot
}
