% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_axgb_result.R
\name{plot_pareto_front_projections}
\alias{plot_pareto_front_projections}
\title{Nonconvex pareto front for a given weight range (limiting random projections).}
\usage{
plot_pareto_front_projections(x = NULL, y = NULL, wt_range = c(0, 1),
  plotly = FALSE)
}
\description{
Nonconvex pareto front for a given weight range (limiting random projections).
}
