% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{robustness_measures}
\alias{robustness_measures}
\title{Robustness Measures}
\arguments{
\item{eps}{[`numeric(1)`]\cr
Magnitude of injected noise. Default: 0.01}

\item{n}{[`integer(1)`]\cr
Repliactions for the corruption process. Results are averaged. Default: 1L.}
}
\description{
We implement the following measures:
}
\details{
Noise corruption on full dataset: `robustnoise` \cr
Featurewise corruption: `robustnoiseperfeat` \cr

The following arguments can be controlled via `extra.args`:
}
\examples{
\dontrun{setMeasurePars(robustnoise, eps = 0.05)}
}
