% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{robustnoise}
\alias{robustnoise}
\title{Noise Injection full dataset}
\format{An object of class \code{Measure} of length 10.}
\usage{
robustnoise
}
\arguments{
\item{eps}{[`numeric(1)`]\cr
Magnitude of injected noise. Default: 0.01}

\item{n}{[`integer(1)`]\cr
Repliactions for the corruption process. Results are averaged. Default: 1L.}
}
\description{
Injects noise into the prediction data. This is done as follows:
For `numeric` features we add rnorm(0, diff(range(x)) * eps)
For `factors` we randomly flip eps % to a randomly selected factor level
}
\keyword{datasets}
