% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{robustnoiseperfeat}
\alias{robustnoiseperfeat}
\title{Featurewise Noise Injection}
\format{An object of class \code{Measure} of length 10.}
\usage{
robustnoiseperfeat
}
\arguments{
\item{eps}{[`numeric(1)`]\cr
Magnitude of injected noise. Default: 0.01}

\item{n}{[`integer(1)`]\cr
Repliactions for the corruption process. Results are averaged. Default: 1L.}
}
\description{
Injects noise into each column of the prediction data.
This is done as follows:
For `numeric` features we add rnorm(0, diff(range(x)) * eps)
For `factors` we randomly flip eps % to a randomly selected factor level.
The predictions on original data and on corrupted data are then compared using
accuracy as a measure.
This corruption process is iterated for every feature and averaged.
}
\keyword{datasets}
