% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stopwatch.R
\docType{data}
\name{Stopwatch}
\alias{Stopwatch}
\title{Control runtime of AutoxgboostMC runs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Stopwatch
}
\description{
Control runtime of AutoxgboostMC runs
}
\keyword{datasets}
