% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{xgb.sparsity}
\alias{xgb.sparsity}
\title{Sparsity as percentage of used features}
\format{An object of class \code{Measure} of length 10.}
\usage{
xgb.sparsity
}
\description{
Sparsity as percentage of used features
}
\keyword{datasets}
