% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoxgboostmc.R
\name{autoxgboostmc}
\alias{autoxgboostmc}
\title{Fit and optimize a xgboost model for multiple criteria.}
\usage{
autoxgboostmc(task, measures = NULL, control = NULL,
  iterations = 160L, time.budget = 3600L, par.set = NULL,
  max.nrounds = 10^6, early.stopping.rounds = 10L,
  early.stopping.fraction = 4/5, build.final.model = TRUE,
  design.size = 15L, impact.encoding.boundary = 10L,
  mbo.learner = NULL, nthread = NULL, tune.threshold = TRUE)
}
\arguments{
\item{measures}{[\code{list of \link[mlr]{Measure}}]\cr
Performance measures. If \code{NULL} \code{\link[mlr]{getDefaultMeasure}} is used.}
}
\value{
\code{\link{AutoxgbResult}}
}
\description{
For params and a more in-depth description see `?autoxgboost`.
}
\examples{
\donttest{
iris.task = makeClassifTask(data = iris, target = "Species")
ctrl = makeMBOControl()
ctrl = setMBOControlTermination(ctrl, iters = 1L) #Speed up Tuning by only doing 1 iteration
res = autoxgboostmc(iris.task, control = ctrl, measures = list(mmce))
res
}
}
