% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AxgbOptimizer.R
\docType{data}
\name{AxgbOptimizer}
\alias{AxgbOptimizer}
\title{Abstract Base Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AxgbOptimizer
}
\description{
Abstract Base Class
}
\seealso{
\code{\link{AxgbOptimizerSMBO}}
}
\keyword{datasets}
