% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{fairf1}
\alias{fairf1}
\title{Absolute differences of F1 Scores between groups
See Hardt et al, 2016: https://arxiv.org/pdf/1610.02413.pdf}
\format{An object of class \code{Measure} of length 10.}
\usage{
fairf1
}
\description{
Absolute differences of F1 Scores between groups
See Hardt et al, 2016: https://arxiv.org/pdf/1610.02413.pdf
}
\keyword{datasets}
